Refactor FTP specific logic from Listener Util
